Actual fractal output with Sierpinski's Gasket added

+++ src/coordinates.rs
@@ -1,6 +1,3 @@
-use math::IntPoint;
-use math::RealPoint;
-
 #[derive(Debug, PartialEq)]
 pub struct PlanePoint(pub f64, pub f64);
 
@@ -8,4 +5,4 @@ pub struct PlanePoint(pub f64, pub f64);
 pub struct CameraCoordinates(pub f64, pub f64);
 
 #[derive(Debug, Eq, PartialEq)]
-pub struct CanvasPixel(pub i64, pub i64);
+pub struct CanvasPixel(pub u32, pub u32);
